Move TitleInfo to a different module
